Move `tests.utils.setup_test_homeserver` to `tests.server`
It had no users.

We have just taken the identity of a previous function but don't provide the same
behaviour, so we need to fix this in the next commit...
